requirements.txt
django>=5.2,<6.0

django-cors-headers>=4.9.0
django-environ==0.13.0
djangorestframework>=3.16.1
djoser==2.3.3
social-auth-app-django==5.7.0

pytz==2026.1.post1

Pillow==12.1.1
mutagen==1.47.0

psycopg2==2.9.11

argon2-cffi==25.1.0

djangorestframework-simplejwt>=5.5.1
dj-database-url>=3.1.2
boto3==1.42.68
django-storages==1.14.6

markdown==3.10.2
gunicorn>=25.1.0
whitenoise>=6.12.0

requests==2.32.5

# Google Calendar
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0

# LangGraph
langgraph

# LangChain / AI
langchain-community
langchain-text-splitters
langchain-openai
langchain-core
chromadb
pypdf
python-dotenv

